Move Pipeline to the public filters package
